Labs/Lab13/requirements.txt
langchain-openai
langchain-community
langchain-core
langchain-classic
python-dotenv
faiss-cpu
